packages/utils/src/pollyfill/promise-with-resolvers.ts
// 只在需要的时候声明接口
interface PromiseWithResolvers<T> {
	promise: Promise<T>;
	resolve: (value: T | PromiseLike<T>) => void;
	reject: (reason?: any) => void;
}

if (!Promise.withResolvers) {
	Promise.withResolvers = function <T>(): PromiseWithResolvers<T> {
		let resolve!: (value: T | PromiseLike<T>) => void;
		let reject!: (reason?: any) => void;
		const promise = new Promise<T>((res, rej) => {
			resolve = res;
			reject = rej;
		});
		return { promise, resolve, reject };
	};
}

export {}; // 使这个文件成为一个模块
